make test and patient modules
